# docs/swagger.yaml
definitions:
  domain.UserSub:
    properties:
      ended_at:
        example: "2026-07-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      service_name:
        example: Netflix
        type: string
      service_price:
        example: 990
        type: integer
      started_at:
        example: "2025-07-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655441111
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/subscriptions:
    get:
      description: Возвращает все подписки или подписки конкретного пользователя (если
        передан user_id)
      parameters:
      - description: ID пользователя (UUID)
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            items:
              $ref: '#/definitions/domain.UserSub'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает запись об онлайн-подписке для конкретного пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserSub'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации или некорректный JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет запись об онлайн-подписке для конкретного пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserSub'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное обновление
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации или некорректный JSON
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить запись о подписке
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись по ID подписки (path) и ID пользователя (query)
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: ID пользователя (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ошибка валидации ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаляет запись о подписке
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по её ID (передается в пути)
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные подписки
          schema:
            $ref: '#/definitions/domain.UserSub'
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить одну подписку
      tags:
      - subscriptions
  /api/v1/subscriptions/total:
    get:
      description: 'Считает сумму трат за период. Формат дат: MM-YYYY'
      parameters:
      - description: ID пользователя (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        required: true
        type: string
      - description: Дата начала (01-2025)
        in: query
        name: from
        required: true
        type: string
      - description: Дата окончания (03-2025)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Результат
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Ошибка валидации параметров
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Рассчитать итоговую стоимость
      tags:
      - subscriptions
swagger: "2.0"